Create jade.io.js
Initial attempt at a translator for the JADE.io database. Main issue surrounds trying to obtain the source URL for the case PDF as it is fetched by a HTTP request to a document server and Zotero is unable to intercept the request as simulated by the click(). Therefore, a future task is to understand/emulate the HTTP request from the page so that Zotero may initiate this independently, and create a valid PDF attachment.
For the present moment the functionality only allows for the PDF to be downloaded, and not attached to a reference item.