Create lexisnexis-advance.js
Initial attempt at a Zotero Translator for the Australian Lexis Nexis Advance legal database. Currently appears to work even when used within the Monash login proxy. Currently just the skeleton for the translator, further work needs to be done to retrieve the PDF and to fill out the author, citation etc. info