Update lexisnexis-advance.js
Introduce the use of a HTTP request to the PDF URL. This allows the Zotero Connector to use the login cookie (established at the base/main webpage) when attempting to access the PDF. Attempts to retrieve the PDF without this cookie result in a snapshot of the LexisNexis sign-in page.

TODO: Neaten the retrieval of the casename. The main priority thus far was the PDF retrieval mechanism, now that that has been determined focus may be paid to proper case titles, citations etc.